fix error.h logic
