+++ lib/server/src/listener.cpp
@@ -1,4 +1,5 @@
 #include "listener.h"
+#include "error.h"
 
 Listener::Listener(Router& router, boost::asio::io_context& ioc,
                    boost::asio::ip::tcp::endpoint endpoint,
